+++ front/src/app/[uuid]/check-theme/CheckTheme.module.scss
@@ -25,40 +25,3 @@
     );
   }
 }
-
-.buttons {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 32px;
-
-  @media screen and (min-width: 768px) {
-    gap: 80px;
-  }
-
-  @media screen and (min-width: 1280px) {
-    flex-direction: row;
-  }
-}
-
-/* 戻るボタン */
-.back {
-  position: absolute;
-  bottom: var(--page-padding);
-  left: var(--page-padding);
-
-  @media screen and (min-width: 1024px) {
-    bottom: var(--page-padding-lg);
-    left: var(--page-padding-lg);
-  }
-}
-
-.arrow {
-  position: absolute;
-  top: 0;
-  left: 0;
-  height: 100%;
-  margin-left: 0.25rem;
-}
