+++ front/src/components/elements/BackButton/BackButton.tsx
@@ -1,12 +1,25 @@
+"use client";
+
 import styles from "@/components/elements/BackButton/BackButton.module.scss";
 import { Playlist } from "@/components/ui/tailwind-buttons";
+import { useRouter } from "next/navigation";
+import { useEffect } from "react";
+
 import { IoChevronBack } from "react-icons/io5";
 
-export default function BackButton({ onClick }: { onClick: () => void }) {
+export default function BackButton({ path }: { path: string }) {
+  const router = useRouter();
+  const handleBack = () => {
+    router.push(path);
+  };
+  useEffect(() => {
+    console.log(path);
+  }, [path]);
+
   return (
     <div className={styles.back}>
       <IoChevronBack className={styles.arrow} />
-      <Playlist onClick={onClick} type="button">
+      <Playlist onClick={handleBack} type="button">
         もどる
       </Playlist>
     </div>
