+++ front/src/features/check-theme/components/Buttons/Buttons.module.scss
@@ -1,0 +1,16 @@
+.buttons {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 32px;
+
+  @media screen and (min-width: 768px) {
+    gap: 80px;
+  }
+
+  @media screen and (min-width: 1280px) {
+    flex-direction: row;
+  }
+}
