+++ front/src/features/check-theme/components/index.ts
@@ -1,0 +1,3 @@
+import Buttons from "@/features/check-theme/components/Buttons/Buttons";
+
+export { Buttons };
